Run Shiny app validation tools in your project

Run all specified tools and requirements to validate Shiny apps project.

audit_app(
  cran = FALSE,
  vignettes = FALSE,
  error_on = "never",
  timeout = 5,
  headless_actions = NULL,
  workers = 5,
  scope = c("manual", "DMC", "POC"),
  output_validation = TRUE,
  coverage = TRUE,
  load_testing = TRUE,
  profile_code = TRUE,
  check_reactivity = TRUE,
  flow = TRUE
)

Arguments

cran

Whether to apply as CRAN check. Defaults to FALSE.

vignettes

Whether to build vignettes. Defaults to FALSE.

error_on

When to raise an error. Possible choices: c("never", "error", "warning", "note"). Defaults to never.

timeout

Time to wait after starting the subprocess (s). Useful is you know how much time the app takes to load.

headless_actions

Custom code passed as a string to manipulate the app with headless web browser, for instance "headless_app$set_inputs(obs = 200); headless_app$run_js('1+1');". If NULL, the app will undergo a classic random Monkey test session.

workers

Number of workers for shinycannon. Default to 5.

scope

Project scope. Accepted values c("manual", "DMC", "POC").

output_validation

Whether to compare output snapshots for plots and htmlwidgets. Default to TRUE.

coverage

Whether to perform coverage report. Default to TRUE.

load_testing

Whether to perform load test. Default to TRUE.

profile_code

Whether to profile R code. Default to TRUE.

check_reactivity

Whether to check reactivity log. Default to TRUE.

flow

Whether to display project overview. Default to TRUE.